Add net.mindview.util package, Range and Print tool classes.

+++ operations/EqualsMethod2.java
@@ -10,7 +10,7 @@ public class EqualsMethod2{
 		Value v1 = new Value();
 		Value v2 = new Value();
 		v1.i = v2.i = 100;
-		System.out.println(v1.equals(v2.intValue()));
+		System.out.println(v1.equals(v2));
 		System.out.println(v1 == v2);
 	}
 }
